# startUp/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AiEd Learning Assistant",
    "version": "1.0.0",
    "description": "AiEd Learning Assistant API Documentation"
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "days": {
        "type": "object",
        "required": ["title", "date", "id", "day", "index", "read"],
        "properties": {
          "title": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "day": {
            "type": "string"
          },
          "index": {
            "type": "string"
          },
          "read": {
            "type": "string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/{lang}/quarters/{quarter_id}/lessons/{lesson_id}/days": {
      "post": {
        "summary": "Create Day",
        "description": "Create Day",
        "tags": ["Days"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/days"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "get": {
        "summary": "Get all Days",
        "description": "Get all Days by language, quarter and lesson",
        "tags": ["Days"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "en"
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2023_03"
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "01"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/days"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/{lang}/quarters/{quarter_id}/lessons/{lesson_id}/days/{day_id}": {
      "get": {
        "summary": "Get Specific Day",
        "description": "Get Specific Day",
        "tags": ["Days"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "en"
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "2023_03"
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "01"
          },
          {
            "name": "day_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "01"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/days"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Day",
        "description": "Update Day",
        "tags": ["Days"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "day_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/days"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/days"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Day",
        "description": "Delete Day",
        "tags": ["Days"],
        "parameters": [
          {
            "name": "lang",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "quarter_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "day_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  }
}